requirements.txt
tensorflow
scikit-learn
dvc
pandas
numpy
python-box
flask
Pillow
-e .